Comment out serde directives Move serialization functions into a seperate module
